Add files via upload
In this project, I analysed the nutrition dataset of the United States Department of Agriculture(USDA). I provided answers to the questions and I provided insights that would help the department to understand nutrients available to the American people. The data for this project was gotten from Harry Moreno's GitHub repository (https://github.com/morenoh149/postgresDBSamples/tree/master/usda-r18-1.0) and loaded to a Postgres database.